debezium/dbz#2217 Map MySQL and MariaDB REAL columns to DOUBLE
MySQL and MariaDB treat REAL as a synonym for DOUBLE unless the
REAL_AS_FLOAT sql_mode is enabled; the server rewrites the column type
at DDL time and stores and replicates such columns as 8-byte doubles.

The ANTLR DDL parsers resolved the REAL keyword to java.sql.Types.REAL
(single precision in JDBC terms), so tables whose DDL was captured from
the binlog were emitted with a FLOAT32 schema and their values narrowed
via Float#floatValue(), silently losing precision. The snapshot path
parses the SHOW CREATE TABLE output, where the server already reports
the column as DOUBLE, producing a FLOAT64 schema - the same column
flipped between schemas depending on the capture path.

Resolve REAL to Types.DOUBLE in all three parsers so the streaming path
matches the snapshot path and no precision is lost. Default values for
REAL columns are now emitted as Double accordingly.

+++ debezium-connector-mariadb/src/main/java/io/debezium/connector/mariadb/antlr/MariaDbAntlrDdlParser.java
@@ -163,7 +163,10 @@ public class MariaDbAntlrDdlParser extends AntlrDdlParser<MariaDBLexer, MariaDBP
                         .setSuffixTokens(MariaDBParser.SIGNED, MariaDBParser.UNSIGNED, MariaDBParser.ZEROFILL),
                 new DataTypeResolver.DataTypeEntry(Types.BIGINT, MariaDBParser.INT8)
                         .setSuffixTokens(MariaDBParser.SIGNED, MariaDBParser.UNSIGNED, MariaDBParser.ZEROFILL),
-                new DataTypeResolver.DataTypeEntry(Types.REAL, MariaDBParser.REAL)
+                // REAL is by default a synonym for DOUBLE: the server stores and replicates such columns
+                // as 8-byte doubles. Only when the REAL_AS_FLOAT sql_mode is enabled is REAL a synonym
+                // for FLOAT instead (debezium/dbz#2217)
+                new DataTypeResolver.DataTypeEntry(Types.DOUBLE, MariaDBParser.REAL)
                         .setSuffixTokens(MariaDBParser.SIGNED, MariaDBParser.UNSIGNED, MariaDBParser.ZEROFILL),
                 new DataTypeResolver.DataTypeEntry(Types.DOUBLE, MariaDBParser.DOUBLE)
                         .setSuffixTokens(MariaDBParser.PRECISION, MariaDBParser.SIGNED, MariaDBParser.UNSIGNED, MariaDBParser.ZEROFILL),
